requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1606234718434/work
aiohttp @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/aiohttp_1637857814678/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
base58==2.1.1
bitarray==1.2.2
blinker==1.4
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1636012319596/work
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1636046173594/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1637858084653/work
click==8.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.10.3
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1636022092101/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1637648810489/work
cycler==0.11.0
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
flatbuffers==2.0
fonttools==4.28.2
frozenlist @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/frozenlist_1637854040629/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib==0.4.1
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/grpcio_1637592348692/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1609497507927/work
hexbytes==0.2.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
importlib-metadata @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/importlib-metadata_1633330991090/work
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
keras==2.7.0
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.2
libclang==12.0.0
lru-dict==1.1.7
Markdown @ file:///Users/ktietz/demo/mc3/conda-bld/markdown_1629712599908/work
matplotlib==3.5.0
multiaddr==0.0.9
multidict @ file:///Users/ktietz/demo/mc3/conda-bld/multidict_1628859257161/work
netaddr==0.8.0
numpy @ file:///Users/ktietz/demo/mc3/conda-bld/numpy_1629783711335/work
oauthlib==3.1.0
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==21.3
parsimonious==0.8.1
Pillow==8.4.0
protobuf==3.17.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1636020570107/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pycryptodome==3.14.1
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1610910308735/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing==3.0.6
pyrsistent==0.18.1
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1635862736740/work
python-dateutil==2.8.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1637771257551/work
requests-oauthlib==1.3.0
rlp==2.0.1
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1636009262461/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1637806880936/work
setuptools-scm==6.3.2
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///Users/runner/miniforge3/conda-bld/tensorboard-data-server_1636045914528/work/tensorboard_data_server-0.6.0-py3-none-macosx_11_0_arm64.whl
tensorboard-plugin-wit==1.6.0
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
termcolor==1.1.0
tomli==1.2.2
toolz==0.11.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1632350318291/work
varint==1.0.2
web3==5.28.0
websockets==9.1
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1635505089296/work
wrapt @ file:///Users/ktietz/demo/mc3/conda-bld/wrapt_1629147080301/work
yarl @ file:///Users/ktietz/demo/mc3/conda-bld/yarl_1628859370630/work
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
